+++ Pfhorge Source/Data Objects/PhAbstractName.m
@@ -37,7 +37,7 @@
 {
     NSData *theData = [NSArchiver archivedDataWithRootObject:myName];
     long length = [theData length];
-    ExportLong(length);
+    ExportLong((int)length);
     [myData appendData:theData];
     [super superClassExportWithIndex:index selfData:myData futureData:futureData mainObjects:mainObjs];
 }
